Added ability to edit meetings.
Added an edit button to meetings so that they can be edited by the original author after they have been submitted.
The details text area is now a WYSIWYG editor to make it easier to enter details.
Sending of verification emails is now configurable in config.js, mainly to easily disable in dev environments.
Added javascript to configure the carousel so that all slides are the same height so content isn't moving up and down all the time.