Attempt to fix server code coverage on travis
I'm not sure why the code coverage on travis for the server quit
working.  If I run npm test locally everything works as expected.  If I
copy the exact same command and run it individually I get the same error
travis was seeing.  This change fixes that error, so hopefully it will
work on travis now.