Add option to send email when scheduling meeting.
Now when a meeting suggestion is selected and a date is set there is an option to send out an email to all
subscribers with the meeting information.  A new user setting was added to allow users to control if they get this notification.
It will default to the users new meeting suggestion email notification preference.